+++ src/Lego/Field/Concerns/ValueOperator.php
@@ -1,4 +1,4 @@
-<?php namespace Lego\Field\Operators;
+<?php namespace Lego\Field\Concerns;
 
 trait ValueOperator
 {
