+++ src/Lego/Field/Provider/NumberRange.php
@@ -1,7 +1,7 @@
 <?php namespace Lego\Field\Provider;
 
-use Lego\Field\Operators\BetweenFilterTrait;
-use Lego\Field\Operators\FilterOnly;
+use Lego\Field\Concerns\BetweenFilterTrait;
+use Lego\Field\Concerns\FilterOnly;
 
 class NumberRange extends Number
 {
