+++ src/Lego/Foundation/Concerns/InitializeOperator.php
@@ -1,4 +1,4 @@
-<?php namespace Lego\Foundation\Operators;
+<?php namespace Lego\Foundation\Concerns;
 
 trait InitializeOperator
 {
