+++ src/Lego/Widget/Concerns/FieldOperator.php
@@ -1,4 +1,4 @@
-<?php namespace Lego\Widget\Operators;
+<?php namespace Lego\Widget\Concerns;
 
 use Illuminate\Support\Collection;
