+++ src/Lego/Widget/Concerns/RequestOperator.php
@@ -1,4 +1,4 @@
-<?php namespace Lego\Widget\Operators;
+<?php namespace Lego\Widget\Concerns;
 
 use Illuminate\Support\Facades\Request;
